use explicit source for fetching image + add scheme and registry tests
