+++ syft/source/scheme.go
@@ -21,44 +21,39 @@ const (
 	ImageScheme Scheme = "ImageScheme"
 )
 
-func detectScheme(fs afero.Fs, imageDetector sourceDetector, userInput string) (Scheme, string, error) {
+func detectScheme(fs afero.Fs, imageDetector sourceDetector, userInput string) (Scheme, image.Source, string, error) {
 	if strings.HasPrefix(userInput, "dir:") {
 		// blindly trust the user's scheme
 		dirLocation, err := homedir.Expand(strings.TrimPrefix(userInput, "dir:"))
 		if err != nil {
-			return UnknownScheme, "", fmt.Errorf("unable to expand directory path: %w", err)
+			return UnknownScheme, image.UnknownSource, "", fmt.Errorf("unable to expand directory path: %w", err)
 		}
-		return DirectoryScheme, dirLocation, nil
+		return DirectoryScheme, image.UnknownSource, dirLocation, nil
 	}
 
 	// we should attempt to let stereoscope determine what the source is first --but, just because the source is a valid directory
 	// doesn't mean we yet know if it is an OCI layout directory (to be treated as an image) or if it is a generic filesystem directory.
 	source, imageSpec, err := imageDetector(userInput)
 	if err != nil {
-		return UnknownScheme, "", fmt.Errorf("unable to detect the scheme from %q: %w", userInput, err)
+		return UnknownScheme, image.UnknownSource, "", fmt.Errorf("unable to detect the scheme from %q: %w", userInput, err)
 	}
 
-	switch source {
-	case image.OciRegistrySource, image.DockerDaemonSource:
-		// return the original user input, not the image ref. In this way we ensure that stereoscope has enough
-		// information to make the final decision on the image source.
-		return ImageScheme, userInput, nil
-	case image.UnknownSource:
+	if source == image.UnknownSource {
 		dirLocation, err := homedir.Expand(userInput)
 		if err != nil {
-			return UnknownScheme, "", fmt.Errorf("unable to expand potential directory path: %w", err)
+			return UnknownScheme, image.UnknownSource, "", fmt.Errorf("unable to expand potential directory path: %w", err)
 		}
 
 		fileMeta, err := fs.Stat(dirLocation)
 		if err != nil {
-			return UnknownScheme, "", nil
+			return UnknownScheme, source, "", nil
 		}
 
 		if fileMeta.IsDir() {
-			return DirectoryScheme, dirLocation, nil
+			return DirectoryScheme, source, dirLocation, nil
 		}
-		return UnknownScheme, "", nil
+		return UnknownScheme, source, "", nil
 	}
 
-	return ImageScheme, imageSpec, nil
+	return ImageScheme, source, imageSpec, nil
 }
